Move jsonhandler and markdown to logic
