+++ src/embeds/homebrew.py
@@ -10,7 +10,7 @@ from components.items import (
 from components.modals import SimpleModal
 from components.paginated_view import PaginatedLayoutView
 from logic.homebrew import HomebrewData, HomebrewEntry, HomebrewEntryType
-from markdown import MDFile, wrapped_md_table_to_rich_table
+from logic.markdown import MDFile, wrapped_md_table_to_rich_table
 
 
 class HomebrewEmbed(discord.Embed):
